Bumping up the mass of the Shuttle and Star barge, and adding a "Medium Freighter" category
The shuttle, heavy shuttle, and star barge just feel a little too agile compared to the interceptors. Bumping up their mass by 30 drops it down a little, but still very spry. (note that the star barge starts off "OK", but has the engine space to become "great".

The "Medium Freighter" addition is for ease of comparison and making the progression a bit more visible; but I do like the look of it in the shipyard. (In other words, this isn't a permanent change, but I'd like it to be if others agree it useful/beneficial.)