Reducing counts to meet the limits
All instances of removed thrusters, steering, or afterburners will be commented so I can track where they are. Extra stand alone thrusters are just commented out. Ex, if a ship has a large thruster and a small thruster, I'll comment out the small thruster.l

The 'olofez (world-ship) was an oddity, not entirely sure of its role as an HD capable fighter. Clever build, makes me wish that the Korath had a smaller power generator (<=9 tons. Perhaps a solar? Given their proximity to the core, the Korath having at least one solar panel would be appropriate...

Important note: This is not the end state. I'll be going through every one of these to see if the removal of those excess thrusters was meaningful, detrimental, or contrary to lore. Likewise, there is also a general rebalance of all the thrusters and steering systems planned. So don't make any jugements based on things as they stand here. Also, this isn't all of them. There's more to go, this is just the end of the work day for this.