Adding more info sections
As of this commit, we now have six different Draw... sections:
- Name
- Model
- Cost
- Health
- CarryingCapacities
- ManeuverStats

Each of these can be called individually, and if the call to any one of them is commented out, the rest move up to fill the space appropriately.

Note: I am missing the "tons" label on the current mass. Will need to fix that.
- I am also considering removing the two decimanl points on the max speed and thrust related stats. It's really not relevant.