asteroids
The belt goes between the orbits of the two stars - I know that doesn't work from an orbital mechanics perspective, but neither does any circular orbit around the system centre that's not far enough out to treat the two stars as a single mass.

Added more mineables since the belt is bigger and hence more spread out - the density will be very low, but that makes sense. The occasional stray field asteroids do sort of make sense in this system; due to the aforementioned lack of stable circular orbits you'd expect to see stray ones flying all over the place.