Shrinking some human scanners a bit and Remnant changes
Salvage Scanner now does both active scans and asteroid scans (so all the active/offensive stuff), while the Laboratory does all the passive targeting stuff.

Regarding the values on the Lab. I picked an amount that would let human scanners exceed the Laboratory if they filled the same amount of space with just one scanner type; but as a complete package for all the values the Remnant one is significantly better.

Reduced the size of the asteroid scanner by 1 and the tactical scanner down to 2, since it doesn't do as many things anymore.