Removals to go back to as close to original as possible
Basically, walking back changes to get rid of any change I can.

Still generates the crash, but hopefully the fewer changes that exist, the easier to pick out what is causing it.

This adds a few ships to the shipyard, and adds two pylons to the clipper just for testing purposes. Will be removed later.