Alternate solution
Vitalchip noted that the count increases by 2 is to generate pairs of attributes, which the 3 would cause problems with. They noted that there is a missing comma on L947 that would cause it to be 5 items instead of 6, which could also cause problems since the last set wouldn't be a pair.

Tested, and this also solves the crash.