Bunrodea
Honestly, I don't have a good feel for this faction. They seem specialized with lore emphasizing lots of guns and downplaying turrets. Not sure if they'd downplay missiles too. So this basically just converts guns to pylons on the sasumata and tanto, gives pylons to the Kaiken, and then adds the requisite pylons onto those ships that have Swarm pods.