Apply suggestions from code review by ravenshining. Thank you!
This isn't everything, yet. One I want to leave there so it reminds me to do the things that you couldn't suggest on, and the others I'd like to look at the reasons behind the order change. Generally, I'm trying to keep pylons after guns, in the same way that we have turrets after guns. (so that guns are together, then the pylons, then the turrets)

Co-authored-by: Lia <31712538+ravenshining@users.noreply.github.com>